// fe/src/lib/api/generated/model/exitInterviewDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AcademicYearSummary } from './academicYearSummary';
import type { CareerModeValue } from './careerModeValue';
import type { CollegeFeedbackRatingResponse } from './collegeFeedbackRatingResponse';
import type { DelayReasonValue } from './delayReasonValue';
import type { ExitInterviewStatusValue } from './exitInterviewStatusValue';
import type { ProgramCompletionValue } from './programCompletionValue';
import type { ReopenEventResponse } from './reopenEventResponse';
import type { SelfAssessmentRatingResponse } from './selfAssessmentRatingResponse';
import type { SignificantLearningValue } from './significantLearningValue';
import type { StudentSummary } from './studentSummary';
import type { StudyCareerChoiceValue } from './studyCareerChoiceValue';
import type { WorkCareerChoiceValue } from './workCareerChoiceValue';

export interface ExitInterviewDetailResponse {
  academic_year: AcademicYearSummary;
  age?: number | null;
  career_modes?: CareerModeValue[];
  civil_status?: string;
  college_feedback_ratings: CollegeFeedbackRatingResponse[];
  contact_number?: string;
  course?: string;
  created_at: string;
  curriculum_comments?: string;
  dean_comments?: string;
  delay_other?: string;
  delay_reasons?: DelayReasonValue[];
  email_address?: string;
  extra_terms_count?: number | null;
  facilities_comments?: string;
  faculty_comments?: string;
  first_submitted_at: string | null;
  guidance_counselor_comments?: string;
  home_address?: string;
  id: string;
  inventory_id: string;
  last_submitted_at: string | null;
  major?: string;
  office_staff_comments?: string;
  program_chair_comments?: string;
  program_completion?: ProgramCompletionValue | null;
  reopen_events: ReopenEventResponse[];
  self_assessment_ratings: SelfAssessmentRatingResponse[];
  significant_learning_experiences?: SignificantLearningValue[];
  significant_learning_other?: string;
  status: ExitInterviewStatusValue;
  student: StudentSummary;
  student_name?: string;
  study_choices?: StudyCareerChoiceValue[];
  suggestions_recommendations?: string;
  updated_at: string;
  work_choices?: WorkCareerChoiceValue[];
}