// fe/src/lib/api/generated/model/exitInterviewDraftPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { CareerModeValue } from './careerModeValue';
import type { CollegeFeedbackRatingPayload } from './collegeFeedbackRatingPayload';
import type { DelayReasonValue } from './delayReasonValue';
import type { ProgramCompletionValue } from './programCompletionValue';
import type { SelfAssessmentRatingPayload } from './selfAssessmentRatingPayload';
import type { SignificantLearningValue } from './significantLearningValue';
import type { StudyCareerChoiceValue } from './studyCareerChoiceValue';
import type { WorkCareerChoiceValue } from './workCareerChoiceValue';

export interface ExitInterviewDraftPayload {
  age?: number | null;
  career_modes?: CareerModeValue[];
  civil_status?: string;
  college_feedback_ratings?: CollegeFeedbackRatingPayload[];
  contact_number?: string;
  course?: string;
  curriculum_comments?: string;
  dean_comments?: string;
  delay_other?: string;
  delay_reasons?: DelayReasonValue[];
  email_address?: string;
  extra_terms_count?: number | null;
  facilities_comments?: string;
  faculty_comments?: string;
  guidance_counselor_comments?: string;
  home_address?: string;
  major?: string;
  office_staff_comments?: string;
  program_chair_comments?: string;
  program_completion?: ProgramCompletionValue | null;
  self_assessment_ratings?: SelfAssessmentRatingPayload[];
  significant_learning_experiences?: SignificantLearningValue[];
  significant_learning_other?: string;
  student_name?: string;
  study_choices?: StudyCareerChoiceValue[];
  suggestions_recommendations?: string;
  work_choices?: WorkCareerChoiceValue[];
}