// fe/src/lib/api/generated/model/selfAssessmentItemValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type SelfAssessmentItemValue = typeof SelfAssessmentItemValue[keyof typeof SelfAssessmentItemValue];


export const SelfAssessmentItemValue = {
  PRIDE_CONFIDENCE_CNSC: 'PRIDE_CONFIDENCE_CNSC',
  ACADEMIC_RECREATION_BALANCE: 'ACADEMIC_RECREATION_BALANCE',
  HOLISTIC_PERSONAL_WELL_BEING: 'HOLISTIC_PERSONAL_WELL_BEING',
  INTEGRATE_KNOWLEDGE_EXPERIENCE: 'INTEGRATE_KNOWLEDGE_EXPERIENCE',
  CAREER_GOAL_CLARITY: 'CAREER_GOAL_CLARITY',
  SELF_ESTEEM: 'SELF_ESTEEM',
  SELF_AWARENESS: 'SELF_AWARENESS',
  COPE_WITH_PRESSURES: 'COPE_WITH_PRESSURES',
  DEAL_WITH_DIFFERENT_WALKS: 'DEAL_WITH_DIFFERENT_WALKS',
  LEADERSHIP: 'LEADERSHIP',
  COMMUNICATION_SKILLS: 'COMMUNICATION_SKILLS',
  CIVIC_MINDEDNESS: 'CIVIC_MINDEDNESS',
  INITIATIVE: 'INITIATIVE',
  DECISION_MAKING: 'DECISION_MAKING',
  RELATIONSHIP_WITH_GOD: 'RELATIONSHIP_WITH_GOD',
} as const;