// fe/src/lib/api/generated/model/careerModeValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CareerModeValue = typeof CareerModeValue[keyof typeof CareerModeValue];


export const CareerModeValue = {
  WORK: 'WORK',
  STUDY: 'STUDY',
} as const;