// fe/src/lib/api/generated/model/delayReasonValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type DelayReasonValue = typeof DelayReasonValue[keyof typeof DelayReasonValue];


export const DelayReasonValue = {
  TRANSFEREE: 'TRANSFEREE',
  ACADEMIC_FAILURES: 'ACADEMIC_FAILURES',
  OTHER: 'OTHER',
} as const;