// fe/src/lib/api/generated/model/programCompletionValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ProgramCompletionValue = typeof ProgramCompletionValue[keyof typeof ProgramCompletionValue];


export const ProgramCompletionValue = {
  ACCORDING_TO_SCHEDULE: 'ACCORDING_TO_SCHEDULE',
  WITH_SOME_DELAY: 'WITH_SOME_DELAY',
} as const;